+++ src/constants/checkArr.js
@@ -1,0 +1,7 @@
+export const checkArr = [
+  { name: 'Услуга 1 ', id: 1 },
+  { name: 'Услуга 2 ', id: 2 },
+  { name: 'Услуга 3 ', id: 3 },
+  { name: 'Услуга 4 ', id: 4 },
+  { name: 'Услуга 5 ', id: 5 },
+];
