+++ src/constants/constantsValue.js
@@ -7,18 +7,3 @@ export const contactTelephoneAndEmail = {
   telephone: '',
   email: '',
 };
-
-export const checkArr = [
-  { name: 'Услуга 1 ', id: 1 },
-  { name: 'Услуга 2 ', id: 2 },
-  { name: 'Услуга 3 ', id: 3 },
-  { name: 'Услуга 4 ', id: 4 },
-  { name: 'Услуга 5 ', id: 5 },
-];
-
-export const tabArr = [
-  { name: 'Tab 1', id: 1 },
-  { name: 'Tab 2', id: 2 },
-  { name: 'Tab 3', id: 3 },
-  { name: 'Tab 4', id: 4 },
-];
