+++ src/constants/tabArr.js
@@ -1,0 +1,6 @@
+export const tabArr = [
+  { name: 'Tab 1', id: 1 },
+  { name: 'Tab 2', id: 2 },
+  { name: 'Tab 3', id: 3 },
+  { name: 'Tab 4', id: 4 },
+];
